Feature - add helpers to optimise fetching and parsing multiple metadata accounts (#31)
* Feature - add helpers to optimise fetching and parsing
multiple metadata accounts

I am working on an app which parses many transactions using codama and as part of this I need to fetch IDLs for all programs in all instructions. As the only published method for fetching and parsing program-metadata content only supports single calls this results in many getAccountInfo calls which can exceed RPC rate limits for some suppliers.

The addition of fetchAndParseAllMetadataContent which uses getMultipleAccounts to prefetch all PDA accounts, and additionally fetches all external accounts in a total of 2 RPC calls instead of potentially hundreds.

* dont assert all accounts exist, we may want to call for a batch of programs, not all will have metadata

* Fixed lint and addressed points raised in PR